// src/app/interfaces/table-widget-info.ts
export interface TableWidgetInfo {
    enableColumnResizing: boolean;
    enableFiltering: boolean;
    enableGridMenu: boolean;
    enableSorting: boolean;
    fastWatch: boolean;
    rowHeight: number;
    showColumnFooter: boolean;
    showGridFooter: boolean;
    tableHeight: number;
    columnDefs: any[];
    tableType: number;
}